🔷 update mods, quests and scripts
🛠️ improve 'update and sync' script
➕ add some trofers
⬆️ update Allthemodium
⬆️ update Armor Statues
⬆️ update Better Mods Button
⬆️ update Crafting Tweaks
⬆️ update Create Enchantment Industry
⬆️ update Create to Modrinth release
⬆️ update Cupboard
⬆️ update EMI
⬆️ update Engineer's Decor
⬆️ update EvilCraft
⬆️ update FTB Quests
⬆️ update Fusion Connected Textures
⬆️ update Kotlin Forge
⬆️ update Lootr
⬆️ update Macaw's Bridges
⬆️ update Reliquary
⬆️ update Resourceful Lib
⬆️ update Sophisticated Core
⬆️ update SuperMartijin642's Config Lib
⬆️ update SuperMartijin642's Core Lib
⬆️ update Torchmaster
➖ remove Rubidium/Magnesium Extras (replaced with Rubidium Extra)
➕ add Betther Third Person
➕ add Chipped
➕ add ExtendedAE
➕ add Rechilesed
➕ add Rechilesed Create
➕ add JER Integration
➕ add OkZoomer
➕ add Resourceful Lootbags
➕ add Rubidium Extra

+++ update_and_sync.ps1
@@ -1,17 +1,20 @@
-packwiz refresh --pack-file ".\src\pack.toml" --meta-folder-base ".\src"
+packwiz refresh --pack-file ".\src\pack.toml" --meta-folder-base ".\src" | Out-Null
 Write-Host "Updating mods" -ForegroundColor Yellow
 packwiz update --all --pack-file ".\src\pack.toml" --meta-folder-base ".\src"
 Write-Host "Update finished" -ForegroundColor Green
 
-Remove-Item -Path '.\dist' -Recurse
+if (Test-Path '.\dist') {
+  Remove-Item -Path '.\dist' -Recurse
+}
 
 $distpaths = @(
   '.\dist\client', '.\dist\server',
   '.\dist\client\defaultconfigs', '.\dist\server\defaultconfigs'
 )
+
 foreach ($path in $distpaths) {
   if (-Not (Test-Path $path)) {
-    mkdir -Path $path
+    mkdir -Path $path | Out-Null
   }
 }
 
@@ -24,16 +27,16 @@ Copy-Item -Path ".\src\*" -Destination ".\dist\server" -Recurse -Force -Exclude 
 Copy-Item -Path ".\src\icon.png" -Destination ".\dist\server\server-icon.png" -Force
 Copy-Item -Path ".\src\config\*" -Destination ".\dist\server\defaultconfigs" -Recurse -Force
 
-Write-Host "Removing clientside-only mods from server dist"
+Write-Host "Removing clientside-only mods from server dist" -ForegroundColor Red
 
 $regex = ''
 foreach ($mod in Get-Content -Path ".\clientside-only.txt") {
-  if ($mod -match $regex) {
-    packwiz remove $mod --pack-file ".\dist\server\pack.toml" --meta-folder-base ".\dist\server"
+  if ($mod -match $regex) { 
+    packwiz remove $mod --pack-file ".\dist\server\pack.toml" --meta-folder-base ".\dist\server" | Out-Null
   }
 }
 
-packwiz refresh --pack-file ".\dist\client\pack.toml" --meta-folder-base ".\dist\client"
-packwiz refresh --pack-file ".\dist\server\pack.toml" --meta-folder-base ".\dist\server"
+packwiz refresh --pack-file ".\dist\client\pack.toml" --meta-folder-base ".\dist\client" | Out-Null
+packwiz refresh --pack-file ".\dist\server\pack.toml" --meta-folder-base ".\dist\server" | Out-Null
 
-Write-Host "Sync finished" -ForegroundColor Yellow
+Write-Host "Sync finished" -ForegroundColor Green
